requirements.txt
openai-whisper
pyspellchecker